gh-630 fix loading
